+++ IsIdentifiable/Options/IsIdentifiableBaseOptions.cs
@@ -150,6 +150,12 @@ public class IsIdentifiableBaseOptions : ITargetsFileOptions
     public int? MaxValidationCacheSize { get; set; } = MaxValidationCacheSizeDefault;
 
     /// <summary>
+    /// Optional.  Set to a stop processing after x records e.g. only evalute top 1000 records of a table/file.  Currently only supported for csv/database
+    /// </summary>
+    [Option(HelpText = "Optional.  Set to stop processing after x records e.g. only evalute top 1000 records of a table/file.  Currently only supported for csv/database", Default = -1)]
+    public int Top { get; set; } = -1;
+
+    /// <summary>
     /// Default for <see cref="MaxValidationCacheSize"/>
     /// </summary>
     public const int MaxValidationCacheSizeDefault = 1_000_000;
@@ -171,7 +177,6 @@ public class IsIdentifiableBaseOptions : ITargetsFileOptions
     /// <see cref="GetTargetName"/> and so not respect this property.</para>
     /// </summary>
     public string TargetName { get; set; } = TargetNameDefault;
-
 
     /// <summary>
     /// Location of database connection strings file (for issuing UPDATE statements)
@@ -274,6 +279,10 @@ public class IsIdentifiableBaseOptions : ITargetsFileOptions
 
         if (TargetsFile == IsIdentifiableReviewerOptions.TargetsFileDefault && !string.IsNullOrWhiteSpace(globalOpts.TargetsFile))
             TargetsFile = globalOpts.TargetsFile;
+            
+        // if global options specifies to only run on x records and we don't have an explicit declaration for that property
+        if (Top <= 0 && globalOpts.Top > 0)
+            Top = globalOpts.Top;
     }
 
     /// <summary>