+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 
 ## Added
 
+- New flag `--top` to only run on a subset of the data available (e.g. `top 1000`).  Currently only supported by relational database and csv runners
 - Added ability to ignore whole columns in reviewer by pressing `Del` on the column and confirming
 - Support for naming servers in `Targets.yaml` for main `ii` binary instead of connection strings (e.g. `-d myserver` for running on relational dbs)
+- You can now pass a directory name to the `file` verb to process all csv files in that directory.
+- Added `-g` option to the `file` verb to process multiple csv files e.g. `**/*.csv`.  This option is only valid when specifying a directory for `-f`
+
 
 ## Fixed